Moved User model to the Models folder
